+++ Windows/QuickLookProtein.Settings/MainWindow.xaml.cs
@@ -97,6 +97,8 @@ public partial class MainWindow : Window
         ShowInfoOverlayCheck.IsChecked        = SettingsStore.GetShowInfoOverlay();
         ShowControlsInPreviewCheck.IsChecked  = SettingsStore.GetShowControlsInPreview();
         OutlineShadingCheck.IsChecked         = SettingsStore.GetOutlineShading();
+        AutoOrientCheck.IsChecked             = SettingsStore.GetAutoOrient();
+        CubeIsosurfaceCheck.IsChecked         = SettingsStore.GetCubeIsosurface();
         CtlShowStickCheck.IsChecked    = SettingsStore.GetCtlShowStick();
         CtlShowLineCheck.IsChecked     = SettingsStore.GetCtlShowLine();
         CtlShowSphereCheck.IsChecked   = SettingsStore.GetCtlShowSphere();
@@ -145,6 +147,8 @@ public partial class MainWindow : Window
         ShowInfoOverlayCheck.Click        += (_, _) => Save(() => SettingsStore.SetShowInfoOverlay(ShowInfoOverlayCheck.IsChecked == true));
         ShowControlsInPreviewCheck.Click  += (_, _) => Save(() => SettingsStore.SetShowControlsInPreview(ShowControlsInPreviewCheck.IsChecked == true));
         OutlineShadingCheck.Click         += (_, _) => Save(() => SettingsStore.SetOutlineShading(OutlineShadingCheck.IsChecked == true));
+        AutoOrientCheck.Click             += (_, _) => Save(() => SettingsStore.SetAutoOrient(AutoOrientCheck.IsChecked == true));
+        CubeIsosurfaceCheck.Click         += (_, _) => Save(() => SettingsStore.SetCubeIsosurface(CubeIsosurfaceCheck.IsChecked == true));
         CtlShowStickCheck.Click    += (_, _) => Save(() => SettingsStore.SetCtlShowStick(CtlShowStickCheck.IsChecked == true));
         CtlShowLineCheck.Click     += (_, _) => Save(() => SettingsStore.SetCtlShowLine(CtlShowLineCheck.IsChecked == true));
         CtlShowSphereCheck.Click   += (_, _) => Save(() => SettingsStore.SetCtlShowSphere(CtlShowSphereCheck.IsChecked == true));
@@ -506,6 +510,8 @@ public partial class MainWindow : Window
             .Replace("{CTL_SHOW_LABELCA}",       QuickLookProtein.Shared.SettingsStore.GetCtlShowLabelCA()  ? "true" : "false")
             .Replace("{CTL_SHOW_RECENTER}",      QuickLookProtein.Shared.SettingsStore.GetCtlShowRecenter() ? "true" : "false")
             .Replace("{OUTLINE_SHADING}",        QuickLookProtein.Shared.SettingsStore.GetOutlineShading()  ? "true" : "false")
+            .Replace("{AUTO_ORIENT}",            QuickLookProtein.Shared.SettingsStore.GetAutoOrient()      ? "true" : "false")
+            .Replace("{CUBE_ISOSURFACE}",        QuickLookProtein.Shared.SettingsStore.GetCubeIsosurface()  ? "true" : "false")
             .Replace("{PDB_TITLE}",              "")
             .Replace("{EXTRA_MODELS_JSON}",      "[]")
             .Replace("{EXTRA_MODELS_HTML}",      "")